pyproject.toml
[project]
name = "emoticare"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "google-api-python-client>=2.150.0",
    "gtts>=2.5.4",
    "h5py>=3.12.0",
    "langchain-core>=0.3.10",
    "langchain-groq>=0.2.0",
    "langgraph>=0.2.40",
    "numpy>=1.26.4",
    "ollama>=0.3.3",
    "opencv-python-headless>=4.10.0.84",
    "pyaudio>=0.2.14",
    "python-dotenv>=1.0.1",
    "speechrecognition>=3.10.4",
    "streamlit>=1.40.0",
    "streamlit-webrtc>=0.47.0",
    "tensorflow>=2.17.0",
]
